// Backend/CUD-student/functions/src/interfaces/ApiStockResponse.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface ApiStockResponse {
    meta: Meta;
    body: Body;
}
export interface Body {
    maxAge: number;
    currentPrice: CurrentPrice;
    targetHighPrice: CurrentPrice;
    targetLowPrice: CurrentPrice;
    targetMeanPrice: CurrentPrice;
    targetMedianPrice: CurrentPrice;
    recommendationMean: CurrentPrice;
    recommendationKey: string;
    numberOfAnalystOpinions: Ebitda;
    totalCash: Ebitda;
    totalCashPerShare: CurrentPrice;
    ebitda: Ebitda;
    totalDebt: Ebitda;
    quickRatio: CurrentPrice;
    currentRatio: CurrentPrice;
    totalRevenue: Ebitda;
    debtToEquity: CurrentPrice;
    revenuePerShare: CurrentPrice;
    returnOnAssets: CurrentPrice;
    returnOnEquity: CurrentPrice;
    grossProfits: any[];
    freeCashflow: Ebitda;
    operatingCashflow: Ebitda;
    earningsGrowth: CurrentPrice;
    revenueGrowth: CurrentPrice;
    grossMargins: CurrentPrice;
    ebitdaMargins: CurrentPrice;
    operatingMargins: CurrentPrice;
    profitMargins: CurrentPrice;
    financialCurrency: string;
}

export interface CurrentPrice {
    raw: number;
    fmt: string;
}

export interface Ebitda {
    raw: number;
    fmt: string;
    longFmt: string;
}

export interface Meta {
    version: string;
    status: number;
    copywrite: string;
    symbol: string;
    processedTime: string;
    modules: string;
}